test/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
openai
rouge
python-dotenv
openai